\name{xform}
\alias{xform}
\title{Transforms a data frame with a single sensor type}
\usage{
xform(df)
}
\arguments{
  \item{df}{a data frame}
}
\value{
a data frame where the value column has been transformed
back into native units for that sensor type.
}
\description{
Transforms a data frame with a single sensor type
}

